Provide a way to not flatten on Node 6 and 8
When generating for node 4, we have --flatten to force a flattened
node_modules hierarchy, but on node 6 and no 8 there was no way do
force granular hierarchy.

Introduce --no-flatten command-line parameter to do old-npm style
installation on node 6 and 8.

The main motivation for this is that there is currently a bug in the
flattening algorithm, and this can serve as a workaround until that
has been identified and removed.