Convert CRLF for script bin files
After we started generating our own bin links, we did not provide the
line-ending normalization that npm did. This corrects that to have
similar behavior to before for packages with CRLF line-endings.

This also has a drive-by change to remove the unnecessary
`patchShebangs` due to the `patchShebangsAuto` hook running and finding
these files just fine now.